% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflation_adj.R
\name{inflation_adj}
\alias{inflation_adj}
\title{Adjust for Inflation}
\usage{
inflation_adj(amt, from, to)
}
\arguments{
\item{amt}{A numeric value that represents a dollar value.}

\item{from}{The year of the amt.}

\item{to}{The year you want to adjust the price to.}
}
\description{
Adjust a price or amount for inflation
}
